Fixes dynamic simulation parameters deps cycles
